Restructure
Creating a new file structure to replicate the Tauri guidelines
