+++ .gitignore
@@ -1,7 +1,7 @@
 # See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
 
 # Trunk
-src-ui/dist/
+dist/
 
 # testing
 coverage
