+++ .taurignore
@@ -1,0 +1,3 @@
+/src
+/public
+/Cargo.toml
