+++ index.html
@@ -9,6 +9,7 @@
     <link data-trunk rel="css" href="styles/refresh.module.css" />
     <link data-trunk rel="css" href="styles/langpicker.module.css" />
     <link data-trunk rel="copy-dir" href="public/images"/>
+    <link data-trunk rel="rust" data-wasm-opt="z" />
   </head>
   <body></body>
 </html>
