+++ src-tauri/tauri.conf.json
@@ -3,7 +3,7 @@
     "beforeDevCommand": "trunk serve",
     "devUrl": "http://localhost:1420",
     "beforeBuildCommand": "trunk build",
-    "frontendDist": "../src-ui/dist"
+    "frontendDist": "../dist"
   },
   "bundle": {
     "active": true,
