--- src-ui/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "src-ui"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-console_error_panic_hook = "0.1.7"
-leptos = "0.6.12"
-leptos_router = "0.6.12"
-leptos_sweetalert = "1.0.2"
-serde = { version = "1.0.203", features = ["derive"] }
