--- src-ui/src/pages/index.rs
@@ -1,11 +1,0 @@
-use leptos::*;
-
-#[component]
-pub fn Index() -> impl IntoView {
-    view! {
-        <div>
-            <h1>{"Hello, World!"}</h1>
-            <p>{"This is a simple example of a Leptos app."}</p>
-        </div>
-    }
-}
