+++ src/pages/infected.rs
